[skip] Prevented leaking connections to the Logeater database {{err:28984}} (30m)
Houston has a mechanism whereby connections to _its_ database are managed properly when threaded by the task system; however, the same is not true for connections to the Logeater database. Therefore, we have to do it manually, in this case by wrapping the function in the `with_connection` block. I _think_ this should prevent us leaking connections and needing a restart every 6 days (after we've burned through our 5 pool connections).