[skip] Tied action that posts links to alerts to `github:pull:synchronize` and `github:pull:open` triggers (10m)
I think what was happening before is that the sequence was:
- a PR is updated with new commits
- new commits are created
- the PR associates the commits with itself

This means that when the `commit:create` trigger was firing, the commit did not yet have a PR associated with itself, so it couldn't post a comment. While this new version has to process all the commits every time instead of just new commits, it should still be pretty efficient, only querying alerts once, querying comments once (as before), and then whatever posting needs to be done to actually post the comments.