[skip] Bumped Ruby version to 2.4.6 (1m)
When deploying, Houston does _not_ read from this file. This file is _only_ for local development and should be kept in-sync with what's been configured on the server. The server configuration has to happen in two places: in RVM itself and in the nginx configuration for Houston:

For RVM:
```bash
rvm --default use 2.4.6
```

For nginx/Passenger:
```bash
# /etc/nginx/sites-enabled/houston
passenger_ruby /home/deploy/.rvm/gems/ruby-2.4.6/wrappers/ruby
```